requirements.txt
absl-py==0.7.0
apache-airflow==1.10.9
cleverhans
h5py
keras
lucid
networkx
numpy
pillow
requests
psycopg2-binary
scipy
tensorflow==1.15
tqdm
